Ikke lagre aktoer ved endring av kontaktopplysninger,
