+++ src/soknad-komponenter/sokersFullmektig/sokersfullmektig.js
@@ -20,7 +20,7 @@ const Fullmektig = ({ fullmektig, lagreFullmektig }) => {
   const [visLeggTilKnapp, settVisLeggTilKnapp] = useState(true);
   const toggleVisLeggTilKnapp = () => settVisLeggTilKnapp(!visLeggTilKnapp);
 
-  const [representererKode, settRepresentererKode] = useState(MKV.Koder.representerer.BRUKER);
+  const vedRolleEndring = event => lagreFullmektig(event.target.value, fullmektig.org.orgnr);
 
   return (
     <Nav.Row>
@@ -35,7 +35,7 @@ const Fullmektig = ({ fullmektig, lagreFullmektig }) => {
             <PostAdresse postadresse={fullmektig.org.postadresse} />
           </Fragment>
         }
-        <Nav.Fieldset onChange={e => settRepresentererKode(e.target.value)} onBlur={() => lagreFullmektig(representererKode)} legend="Hvem er dette fullmektig for?" >
+        <Nav.Fieldset onChange={vedRolleEndring} legend="Hvem er dette fullmektig for?" >
           <Nav.Radio label="Arbeidsgiver" name="representerer" value={MKV.Koder.representerer.ARBEIDSGIVER} />
           <Nav.Radio label="Arbeidstaker" name="representerer" value={MKV.Koder.representerer.BRUKER} />
           <Nav.Radio label="Både arbeidstaker og arbeidsgiver" name="representerer" value={MKV.Koder.representerer.BEGGE} />
@@ -44,7 +44,6 @@ const Fullmektig = ({ fullmektig, lagreFullmektig }) => {
       </Nav.Column>
       <Nav.Column xs="6">
         <Kontaktopplysninger
-          representererKode={fullmektig.representererKode}
           juridiskOrg={fullmektig.org}
           visLeggTilKnapp={visLeggTilKnapp}
           toggleVisLeggTilKnapp={toggleVisLeggTilKnapp}
@@ -72,6 +71,18 @@ export class SokersFullmektig extends Component {
     this.hentFullmektige();
   }
 
+  lagreFullmektig = (representererKode, orgnr) => {
+    const { lagreAktoer, oppsummering: { saksnummer } } = this.props;
+    lagreAktoer(saksnummer, {
+      aktoerID: null,
+      orgnr,
+      utenlandskPersonID: null,
+      institusjonsID: null,
+      rolleKode: MKV.Koder.aktoersroller.REPRESENTANT,
+      representererKode,
+    });
+  };
+
   hentFullmektige = async () => {
     const { saksnummer } = this.props.oppsummering;
     const { hentAktoer, hentKontaktopplysninger, hentOrg } = this.props;
@@ -92,7 +103,6 @@ export class SokersFullmektig extends Component {
 
   render() {
     const panelIkon = Ikoner.Ferdig;
-    const { lagreAktoer, oppsummering: { saksnummer } } = this.props;
 
     return (
       <div>
@@ -101,7 +111,7 @@ export class SokersFullmektig extends Component {
           ariaTittel="Opplysninger om fullmektig">
           <Nav.Container fluid>
             {this.state.fullmektige.map(fullmektig => (
-              <Fullmektig key={fullmektig.aktoerID} fullmektig={fullmektig} lagreFullmektig={representererKode => lagreAktoer(saksnummer, MKV.Koder.aktoersroller.REPRESENTANT, representererKode)} />
+              <Fullmektig key={fullmektig.aktoerID} fullmektig={fullmektig} lagreFullmektig={this.lagreFullmektig} />
             ))}
           </Nav.Container>
         </Nav.EkspanderbartpanelBase>
@@ -127,7 +137,7 @@ const mapDispatchToProps = () => ({});
 const hentOrg = async orgNr => Api.Organisasjoner.hentOrganisasjon(orgNr);
 const hentAktoer = async (saksnr, rolleKode, representererKode) => Api.Fagsaker.aktoer.hent(saksnr, rolleKode, representererKode);
 const hentKontaktopplysninger = async (saksnummer, juridiskorgnr) => Api.Fagsaker.kontaktopplysninger.hent(saksnummer, juridiskorgnr);
-const lagreAktoer = async (saksnr, rolleKode, representererKode) => Api.Fagsaker.aktoer.send(saksnr, rolleKode, representererKode);
+const lagreAktoer = async (saksnr, data) => Api.Fagsaker.aktoer.send(saksnr, data);
 
 const SokersFullmektigWrapper = props => <SokersFullmektig {...props} lagreAktoer={lagreAktoer} hentAktoer={hentAktoer} hentOrg={hentOrg} hentKontaktopplysninger={hentKontaktopplysninger} />;
